modularize apiClient
