+++ client/src/services/apiClient/constants.js
@@ -1,0 +1,4 @@
+export const API_BASE_URL = "/api";
+export const DEFAULT_TIMEOUT = 30000;
+export const REFRESH_TIMEOUT = 10000;
+export const LOGOUT_TIMEOUT = 5000;
